actualizar calculo de total en cartitems y cartsummary

+++ app/pago/CartSummary.tsx
@@ -136,7 +136,7 @@ const CartSummary = () => {
                     <h2 className='font-semibold text-primary'>Mostrar resúmen de mi pedido</h2>
                     <KeyboardArrowDown className={`text-primary ${summaryOpened ? 'rotate-180' : ''}`}></KeyboardArrowDown>
                 </div>
-                <span className='font-medium text-heading'>${parseInt(total?.toFixed(2)).toLocaleString('es-ES', { minimumFractionDigits: 2})}<span className='text-small text-text-faded ml-6'>USD</span></span>
+                <span className='font-medium text-heading'>${parseFloat(total?.toFixed(2)).toLocaleString('es-ES', { minimumFractionDigits: 2 })}<span className='text-small text-text-faded ml-6'>USD</span></span>
             </div>
             <div className={`transition-all duration-300 ease-linear ${summaryOpened ? 'max-h-[2000px]' : 'max-h-0'} overflow-hidden lg:max-h-full flex flex-col lg:space-y-24`}>
                 <h2 className='hidden lg:block font-medium text-heading leading-body pb-12 lg:mx-24 border-b-custom text-center'>Resúmen del pedido</h2>
